Revert "SOLR-14067: v3 Create /contrib/scripting module with ScriptingUpdateProcessor (#2215)"
This reverts commit cf5db8d6513e0f3e556ab6ee1b9ad3a6472ad2f2.
