+++ solr/core/src/java/org/apache/solr/update/processor/ScriptEngineCustomizer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.scripting.update;
+package org.apache.solr.update.processor;
 
 import javax.script.ScriptEngine;
