+++ solr/core/src/test-files/solr/collection1/conf/cross-compatible.js
