+++ solr/core/src/test-files/solr/collection1/conf/evil.js
