+++ solr/core/src/test-files/solr/collection1/conf/solrconfig-script-updateprocessor.xml
@@ -18,7 +18,7 @@
 -->
 
 <!--
-   Test Config for ScriptUpdateProcessorFactory
+   Test Config that for ScriptUpdateProcessor
 
   -->
 <config>
@@ -29,7 +29,7 @@
   <schemaFactory class="ClassicIndexSchemaFactory"/>
 
   <updateRequestProcessorChain name="force-script-engine" default="true">
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="engine">javascript</str>
       <str name="script">missleading.extension.updateprocessor.js.txt</str>
     </processor>
@@ -42,7 +42,7 @@
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="single-script">
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">trivial.updateprocessor0.js</str>
       <lst name="params">
         <bool name="boolValue">true</bool>
@@ -53,7 +53,7 @@
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="dual-scripts-arr">
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <arr name="script">
         <str>trivial.updateprocessor0.js</str>
         <str>trivial.updateprocessor1.js</str>
@@ -67,7 +67,7 @@
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="dual-scripts-strs">
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">trivial.updateprocessor0.js</str>
       <str name="script">trivial.updateprocessor1.js</str>
       <lst name="params">
@@ -79,22 +79,22 @@
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="conditional-scripts">
-    <!-- multiple scripts,
+    <!-- multiple scripts, 
          test that the first one can conditionally stop execution -->
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">conditional.updateprocessor.js</str>
       <str name="script">addfields.updateprocessor.js</str>
     </processor>
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="conditional-script">
-    <!-- single script, followed by another processor
-         (that happens to be a script).
+    <!-- single script, followed by another processor 
+         (that happens to be a script). 
          test that the first one can conditionally stop execution -->
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">conditional.updateprocessor.js</str>
     </processor>
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">addfields.updateprocessor.js</str>
     </processor>
   </updateRequestProcessorChain>
@@ -112,13 +112,13 @@
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="javascript-compatibility">
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">cross-compatible.js</str>
     </processor>
   </updateRequestProcessorChain>
 
   <updateRequestProcessorChain name="evil">
-    <processor class="org.apache.solr.update.processor.scripting.StatelessScriptUpdateProcessorFactory">
+    <processor class="solr.StatelessScriptUpdateProcessorFactory">
       <str name="script">evil.js</str>
     </processor>
   </updateRequestProcessorChain>
