+++ solr/core/src/test-files/solr/collection1/conf/throw.error.on.add.updateprocessor.js
