+++ solr/core/src/test/org/apache/solr/core/TestBadConfig.java
@@ -16,6 +16,10 @@
  */
 package org.apache.solr.core;
 
+import javax.script.ScriptEngineManager;
+
+import org.junit.Assume;
+
 public class TestBadConfig extends AbstractBadConfigTestBase {
 
   public void testUnsetSysProperty() throws Exception {
@@ -40,13 +44,35 @@ public class TestBadConfig extends AbstractBadConfigTestBase {
   }
 
   public void testUpdateLogButNoVersionField() throws Exception {
-
+    
     System.setProperty("enable.update.log", "true");
     try {
       assertConfigs("solrconfig.xml", "schema12.xml", "_version_");
     } finally {
       System.clearProperty("enable.update.log");
     }
+  }
+
+  public void testBogusScriptEngine() throws Exception {
+    // sanity check
+    Assume.assumeTrue(null == (new ScriptEngineManager()).getEngineByName("giberish"));
+                      
+    assertConfigs("bad-solrconfig-bogus-scriptengine-name.xml",
+                  "schema.xml","giberish");
+  }
+
+  public void testMissingScriptFile() throws Exception {
+    // sanity check
+    Assume.assumeNotNull((new ScriptEngineManager()).getEngineByExtension("js"));
+    assertConfigs("bad-solrconfig-missing-scriptfile.xml",
+                  "schema.xml","a-file-name-that-does-not-exist.js");
+  }
+
+  public void testInvalidScriptFile() throws Exception {
+    // sanity check
+    Assume.assumeNotNull((new ScriptEngineManager()).getEngineByName("javascript"));
+    assertConfigs("bad-solrconfig-invalid-scriptfile.xml",
+                  "schema.xml","currency.xml");
   }
 
   public void testBogusMergePolicy() throws Exception {
@@ -63,7 +89,7 @@ public class TestBadConfig extends AbstractBadConfigTestBase {
     assertConfigs("bad-solrconfig-managed-schema-named-schema.xml.xml",
                   "schema-minimal.xml", "managedSchemaResourceName can't be 'schema.xml'");
   }
-
+  
   public void testUnknownSchemaAttribute() throws Exception {
     assertConfigs("bad-solrconfig-unexpected-schema-attribute.xml", "schema-minimal.xml",
                   "Unexpected arg(s): {bogusParam=bogusValue}");
