+++ solr/core/src/test/org/apache/solr/update/processor/ScriptEngineTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.solr.scripting.update;
+package org.apache.solr.update.processor;
 
 import org.apache.lucene.util.Constants;
