+++ solr/solr-ref-guide/src/configuring-solrconfig-xml.adoc
@@ -8,7 +8,6 @@
     query-settings-in-solrconfig, \
     requestdispatcher-in-solrconfig, \
     update-request-processors, \
-    script-update-processor, \    
     codec-factory
 
 // Licensed to the Apache Software Foundation (ASF) under one
